+++ src/app.ts
@@ -100,7 +100,7 @@ const run = async () => {
 	// TODO make workers configurable optional
 	await mailService.startWorkers();
 
-	const mongoURI = `mongodb://${process.env.MONGO_HOST || 'localhost'}/notification-service`;
+	const mongoURI = `mongodb://${process.env.MONGO_HOST || 'localhost/notification-service'}`;
 	await mongoose.connect(mongoURI, { useNewUrlParser: true, useUnifiedTopology: true });
 
 	// TODO make producer configurable optional